refactor(example): separate basic components user steps screen
